refactor(domains): give transfers, leagues, wishlist and admin a public API
P2.7. MAY_REACH_INSIDE is now EMPTY -- 9 -> 0. Every domain has an index,
and reaching past it is no longer allowlisted anywhere. That register
started at 34.

  transfers  index.ts        LoanStatusDisplay
             index.server.ts getTransfersDataForDivision
  leagues    index.ts        PositionPointsTable
             index.server.ts getAllLeagueStandingsData, getTeamOfTheWeek
  wishlist   index.ts        WishlistButton, WishlistTags
  admin      index.server.ts handleCommitTeamsToFirestore

wishlist needs no server half -- it is backed by local storage, so none of
it touches Firebase, Sheets or process.env. admin needs no client half:
its components are its own dashboard and nothing outside admin should be
rendering them.

Four of the nine call sites were dynamic imports inside function bodies,
invisible to a grep of the import block -- homepage.route.tsx,
team.server.tsx, system-status.service.ts and draft.server.ts. The
architecture test walks dynamic imports, which is why it keeps finding
these; same lesson as the scoring index.

admin/index.server.ts records an inversion rather than fixing one. Admin
orchestrates other domains -- that is the premise of this item -- but
draft/server/draft.server.ts needs handleCommitTeamsToFirestore, which
lives in admin's server actions, so that one dependency runs the WRONG
way: a feature domain reaching into admin. Exposing it makes the reach
legal without making it right, and the file says so in its own docblock.
The underlying problem is the admin <-> draft cycle, already logged in
Found along the way and carried into P2.6.

Not done, and the other half of this item's acceptance:
TRANSITIONAL_PUBLIC_SEGMENTS. 64 cross-domain imports still reach types/
and lib/ directly. The flag is global so it cannot come out gradually --
flipping it turns all 64 into violations at once, and each domain's index
has to re-export the types its consumers use first. The counts and the
overlap with P2.2 (teams/types is 12 of the 64) are recorded in the
backlog.

Verification: type errors 176 unchanged, 346 tests pass, ratchet counts
unchanged, yarn build clean.
