+++ draft/app/admin/server/transfers-admin.server.tsx
@@ -4,7 +4,7 @@ import { CACHE_KEYS } from '../../_shared/lib/cache/cache-config';
 import { dataCache } from '../../_shared/lib/cache/data-cache.service';
 import type { GameWeekData } from '../../_shared/lib/fpl/fpl-types';
 import type { DivisionId, DivisionSheetData } from '../../_shared/types/league-types';
-import { getTransfersDataForDivision } from '../../transfers/server/services/transfers-data.service';
+import { getTransfersDataForDivision } from '../../transfers/index.server';
 import type { TransferAdminOverviewData, TransferRecommendation } from '../../transfers/types/transfer-rule-types';
 
 interface TransfersActionResult {
